Added option to defer full dataset metadata update
In our tests, updating the TIFF comment for each image in the dataset
caused a massive delay at the end of each timepoint, as the current image
was finished and each other OME-TIFF in the dataset was loaded into
memory, updated, and written back out.

This modification allows us to temporarily suspend these updates; we
sacrifice the assurance of up-to-the-minute metadata for better execution
time.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>